Merging apophis' feature/intergrate_starksoft_proxy_client branch for socks 4/5 support (closes: #165)
Renamed ProxyUser to ProxyUsername and ProxyPass to ProxyPassword.
Default the proxy fields to null, no need to waste CPU cycles for useless assignment.
Actually connect to the proxy server and target server.
Re-use our initialized TcpClient object with the timeout settings for the proxy.
Added own ProxyType enum and map it to the Starksoft's one in IrcClient.Connect().
Updated MonoDevelop project files.