Add support for channel modes +I and +e.
Add support for the RFC2811 channel modes +I and +e. Their appearance on the
wire is more or less equal to +b. +e specifies exceptions to ban masks (if a
user matches both a +b and an +e mode, they are not banned), and +I specifies
"automatic invitations" (if a user matches an +I mode on an +i channel, they may
enter without being invited).