Fix race condition between starting PING stopwatch and receiving PONG
If the server responds with a PONG message quickly enough, the stopwatch
swap-out performed after sending the PING message may come into a race
with the stopwatch swap-out performed after receiving a PONG message. Avoid
this by reordering the PING message to be sent after the swap-out has been
performed.